Moved MEDS-DEV lib files into lib folder
