+++ docs/MEDS-DEV/components/Datasets.jsx
@@ -16,7 +16,7 @@ import {
 import { ContentCopy } from '@mui/icons-material';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 
-import { load_MEDS_DEV, DATASETS } from '@site/src/MEDS-DEV/load.js';
+import { load_MEDS_DEV, DATASETS } from '@site/src/lib/load';
 import Dataset from './Dataset';
 //import EntityPage from './EntityTreeMenu';
 import EntityPage from './EntityPage';
